Added UserAttachmentController
attachments_controller.rb:
  - new CRUD/REST

layout_helper.rb:
  - new function humanized_filesize(bytes)

application_file.rb:
  - process styles only if image-file

attachment.rb:
  - adapted to use with controller

signed_in_menu.haml:
  - Add link to 'Your files …'

routes.rb:
  - Add ressource Attachment in root and nested in :user
  - /attachments and /user/:user_id/attachments

attachment_spec:
  - add file_name-delegate

attachments_controller_spec:
  - new specs for new controller

application_helper_spec:
  - test the new function

+++ app/models/attachment.rb
@@ -1,8 +1,8 @@
 class Attachment
   include Mongoid::Document
   embeds_one :application_file, cascade_callbacks: true
-
-
+  accepts_nested_attributes_for :application_file
+  
   def file
     self.application_file.file if self.application_file
   end
@@ -12,7 +12,11 @@ class Attachment
   end
 
   def path
-    self.application_file.file.path
+    self.application_file.file.path if self.application_file
+  end
+
+  def file_name
+    File::basename(self.path) if self.path
   end
 
   def read
@@ -26,7 +30,7 @@ class Attachment
 
 private
   def clean_file
-    application_file.destroy
+    application_file.destroy if application_file
   end
 
 end
